Separating combinator determination and placement into separate steps
